Add files via upload
Regarding your form: I always like when people make their fonts different, and they chose different colored backgrounds. I will recommend though a more orderly grouping and sizing your form to fit your group boxes. Also, I believe it can be done with a list box, but I personally used a datagrid. Unfortunately, because I didn't use a list box, my later comments will be based on if you had used a datagrid rather than a list box. When I ran the program, I was not able to load the form because it threw an error. I do not know if that is on my part, because this code review is new, and I might have downloaded it wrong, but I will get to the error I received in a minute. I did also check to see if when the form was to load, if it would load in the center of the screen, and it was not (on the form properties, under Layout, it's the third to last option labeled "size," you would just click the dropdown and hit "CenterScreen"), definitely remember to check that next time, it's easy to miss! :) 

The error I received was for line 27, "foreach (Product description in results)." This was completely on me because it was flagging for "the database ... cannot be opened because it is version 904," whatever that means. I use 2022, so I assume you've written this in 2019. Nothing seems wrong with that line of code. I do like that instead of using "db" you used "titan."